[gherkin] No duplicate parser error message
gherkin-go handles the "whitespace in tags" problem with that the parser
stated "yes, it is a tag line - but, there is a parser error".
gherkin-java, gherkin-javascript and gherkin-ruby on the other hand
handles the "whitespace in tags" problem with that the parser stated
"there is a parser error - and no, it is not a tag line". The end result
is that the gherkin-java, gherkin-javascript and gherkin-ruby emits the
"whitespace in tags" error twice (once for the search for a Rule and
once for the search for a Scenario), wheras gherkin-go only emits the
error once.

To reconsile this so that all of them fulfills the same error.ndjson
file, change gherkin-java, gherkin-javascript and gherkin-ruby to not
add an error to the list of errors, if the new error gives out the same
error message as an error already in the list.